Resolve a lot of nitpicks in code
