+++ exonum/src/runtime/types.rs
@@ -284,17 +284,23 @@ impl ValidateInput for InstanceSpec {
     }
 }
 
+/// Configuration parameters of the certain service instance.
 #[derive(Debug, Clone, PartialEq, Eq, Hash, ProtobufConvert, Serialize, Deserialize)]
 #[exonum(pb = "schema::runtime::ServiceConfig", crate = "crate")]
 pub struct ServiceConfig {
+    /// Corresponding service instance ID.
     pub instance_id: InstanceId,
+    /// Raw bytes representation of service configuration parameters.
     pub params: Vec<u8>,
 }
 
+/// This message contains one atomic configuration change.
 #[derive(Debug, Clone, PartialEq, Eq, Hash, ProtobufConvert, Serialize, Deserialize)]
 #[exonum(pb = "schema::runtime::ConfigChange", crate = "crate")]
 pub enum ConfigChange {
+    /// New consensus config.
     Consensus(ConsensusConfig),
+    /// New service instance config.
     Service(ServiceConfig),
 }
